0610 header: atol return type (int->long long)

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: dim <dim@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/25 02:33:16 by dim               #+#    #+#             */
-/*   Updated: 2021/06/09 17:08:14 by dim              ###   ########.fr       */
+/*   Updated: 2021/06/10 21:33:14 by dim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef struct		s_list
 }					t_list;
 
 int					ft_atoi(const char *str);
-int					ft_atol(const char *str);
+long long			ft_atol(const char *str);
 void				ft_bzero(void *s, size_t n);
 void				*ft_calloc(size_t n, size_t size);
 int					ft_isalnum(int c);
